\name{cancer.type.all}
\alias{cancer.type.all}
\docType{data}
\title{
Predefined cancer types in LFSPRO.
}
\description{
We classified the cancers into 11 groups according to NCCN Guidelines Version 1.2012 Li-Fraumeni Syndrome criteria. sts: soft tissue sarcoma; ost: osterosarcoma; brain: brain tumor; breast: breast cancer; acc: adrenocortical carcinoma; leukemia: leukemia; lung: lung bronchoalveolar cancer; choroid: choroid plexus carcinoma; other.lfs: other LFS spectrum cancers; non.lfs: non LFS spectrum invasive cancers; benign: benign tumors.
}
\usage{cancer.type.all}
\format{
 "sts"       "ost"       "brain"     "breast"    "acc"       "leukemia"  "lung"      "choroid"  "other.lfs" "non.lfs"   "benign"
}
\seealso{
\code{\link{lfspro.cancer.type}}, \code{\link{cancer.data}}, \code{\link{lfs.cut}}, \code{\link{invasive.cut}}
}
\examples{
cancer.type.all
}
