\name{invasive.cut}
\alias{invasive.cut}
\docType{data}
\title{
Cutoff for malignant cancer
}
\description{
We coded diffrent cancers into numbers. If the code for a tumor is less than invasive.cut, it is malignant cancer otherwise it is benign tumor.
}
\usage{invasive.cut}
\format{
  The format is:
 num 100
}
\examples{
invasive.cut
}
\keyword{datasets}
