\name{lfs.cut}
\alias{lfs.cut}
\docType{data}
\title{
Cutoff for Li-Fraumeni Syndrome spectrum cancer
}
\description{
We coded diffrent cancers into numbers. If the code for a tumor is less than lfs.cut, it is Li-Fraumeni Syndrome spectrum cancer.
}
\usage{lfs.cut}
\format{
  The format is:
 num 50
}

\examples{
lfs.cut
}
\keyword{datasets}
