\name{cancer.data}
\alias{cancer.data}
\docType{data}
\title{
Cancer information data
}
\description{
Cancer information data storing the cancer type and diagnosed date
}
\usage{cancer.data}
\format{
  A data frame with 4 variables.
  \describe{
    \item{\code{fam.id}}{Family id}
    \item{\code{id}}{Individual id}
    \item{\code{cancer.type}}{Cancer type. See \link{LFSpro.cancer.type} for details}
    \item{\code{diag.age}}{The age when the individual was diagnosed with cancer}
  }
}
\seealso{
\code{\link{LFSpro.cancer.type}}, \code{\link{cancer.type.all}}, \code{\link{lfs.cut}}, \code{\link{invasive.cut}}
}
\examples{
cancer.data
}
\keyword{datasets}
