\name{invasive.cut}
\alias{invasive.cut}
\docType{data}
\title{
Cutoff for malignant cancer
}
\description{
We coded different cancer types using numbers. A cancer is considered malignant if its number is less than invasive.cut, and benign otherwise. 
}
\usage{invasive.cut}
\format{
  The format is:
 num 100
}
\examples{
invasive.cut
}

