\name{lfs.cut}
\alias{lfs.cut}
\docType{data}
\title{
Cutoff for Li-Fraumeni Syndrome spectrum cancer
}
\description{
A cancer is considered within LFS spectrum if its number is less than lfs.cut, and non-LFS spectrum otherwise.
}
\usage{lfs.cut}
\format{
  The format is:
 num 50
}

\examples{
lfs.cut
}

