\name{lfspenet.2010}
\alias{lfspenet.2010}
\docType{data}
\title{
Penetrance for TP53 mutations
}
\description{
Penetrance table of three kinds of genotype (wild type, one allele mutation, two allele mutation) of TP53 gene at age from 1 to 110 for both male and female.
}
\usage{lfspenet.2010}
\format{
List. Two component of list indicates the penetrance of male and female (fMX: male, fFX: female). For each component, it is a 110*3 matrix containing the penetrance from age 1 to age 110 for three kinds of genotype: P530 (wild type), P531 (one allele mutation) and P532 (two allele mutation).
}
\references{
Wu, C. C., et al. (2010). Effects of measured susceptibility genes on cancer risk in family studies. \emph{Hum Genet} \bold{127(3)}: 349-357.
}
\examples{
lfspenet.2010
}

